fix custom intensity range remapping to byte, add detectintensity setting, use ushort for intensity (then convert to byte on save),

+++ Readers/IReader.cs
@@ -19,7 +19,7 @@ namespace PointCloudConverter.Readers
         double GetTime();
 
         void Close();
-        byte GetIntensity();
+        ushort GetIntensity();
         byte GetClassification();
         LasHeader GetMetaData(ImportSettings importSettings, int fileIndex);
